Movie JSON Implementation
Movie is now responsible for parsing JSON string responses form the network for the application for its POJO type.
JsonUtil has now been made a public interface that defines the parseFromJson method.
