+++ app/src/main/java/com/example/nebo/popular_movies/data/Movie.java
@@ -2,12 +2,33 @@ package com.example.nebo.popular_movies.data;
 
 import android.os.Parcel;
 import android.os.Parcelable;
+import android.support.annotation.NonNull;
+import android.util.Log;
 
-public class Movie implements Parcelable {
+import com.example.nebo.popular_movies.util.JsonUtils;
+import com.example.nebo.popular_movies.util.MovieURLUtils;
+
+import org.json.JSONArray;
+import org.json.JSONObject;
+
+import java.util.List;
+import java.util.ArrayList;
+
+public class Movie implements Parcelable, JsonUtils<Movie> {
     // Keep count of the total number of movie objects created.
     private static int mMovieCount = 0;
     // Save the instance of the count to the object created.
     private final int mMovieInstance;
+
+    private static final String MOVIE_LIST_RESPONSE_KEY = "results";
+    private static final String MOVIE_ID_KEY = "id";
+    private static final String MOVIE_VOTE_KEY = "vote_average";
+    private static final String MOVIE_TITLE_KEY = "title";
+    private static final String MOVIE_POPULARITY_KEY = "popularity";
+    private static final String MOVIE_POSTER_PATH_KEY = "poster_path";
+    private static final String MOVIE_BACKDROP_PATH_KEY = "backdrop_path";
+    private static final String MOVIE_OVERVIEW_KEY = "overview";
+    private static final String MOVIE_DATE_KEY = "release_date";
 
     // Actual movie attributes.
     private String mTitle;
@@ -135,4 +156,70 @@ public class Movie implements Parcelable {
             return new Movie[size];
         }
     };
+
+    @Override
+    public List<Movie> parseJsonResponse(@NonNull String response) {
+        List<Movie> movies = new ArrayList<Movie>();
+        JSONObject jsonResponse = null;
+        JSONArray jsonArrayOfMovies = null;
+
+        if (response == null) {
+            return movies;
+        }
+
+        try {
+            jsonResponse = new JSONObject(response);
+            jsonArrayOfMovies = jsonResponse.getJSONArray(Movie.MOVIE_LIST_RESPONSE_KEY);
+
+            for (int index = 0; index < jsonArrayOfMovies.length(); index++) {
+                Movie movie = Movie.parseJsonMovie(jsonArrayOfMovies.getJSONObject(index));
+
+                if (movie != null) {
+                    movies.add(movie);
+                }
+            }
+        }
+        catch (org.json.JSONException e) {
+            Log.e("Parsing JSON Response", "Error in parsing the response " + response);
+            e.printStackTrace();
+        }
+
+        return movies;
+    }
+
+    /**
+     * @brief Parse the JSON object for the individual movie attributes.
+     * @param movieJsonObject json Object that must contain the movie keys.
+     * @return A movie object if successful otherwise null.
+     */
+    private static Movie parseJsonMovie(@NonNull JSONObject movieJsonObject) {
+        Movie movie = null;
+
+        if (movieJsonObject == null) {
+            return movie;
+        }
+
+        // All movie attributes from the JSON response are at the same level.
+        int id;
+        String title, overview, backdropPath, posterPath, date;
+        double popularity, vote;
+
+        try {
+            id = movieJsonObject.getInt(Movie.MOVIE_ID_KEY);
+            title = movieJsonObject.getString(Movie.MOVIE_TITLE_KEY);
+            overview = movieJsonObject.getString(Movie.MOVIE_OVERVIEW_KEY);
+            backdropPath = MovieURLUtils.buildImageUrl(movieJsonObject.getString(Movie.MOVIE_BACKDROP_PATH_KEY)).toString();
+            posterPath = MovieURLUtils.buildImageUrl(movieJsonObject.getString(Movie.MOVIE_POSTER_PATH_KEY)).toString();
+            date = movieJsonObject.getString(Movie.MOVIE_DATE_KEY);
+            popularity = movieJsonObject.getDouble(Movie.MOVIE_POPULARITY_KEY);
+            vote = movieJsonObject.getDouble(Movie.MOVIE_VOTE_KEY);
+        }
+        catch (org.json.JSONException e) {
+            Log.e("Parse Movie Json", "Error in parsing a field for a movie.");
+            e.printStackTrace();
+            return movie;
+        }
+
+        return new Movie(title, id, vote, popularity, posterPath, backdropPath, overview, date);
+    }
 }
